Remove master versions of linux-vanilla and u-boot
Remove all instances of linux-vanilla master and u-boot master versions.
meta-linux-mainline provides a better alternative for testing unmodified
and up-to-date kernel versions on our platforms. meta-phytec provides a
recipe for u-boot master, so we do not need this here, too. As a bonus,
linux-yocto-dev may also be used, if Yocto-specific modifications are
needed on the latest Linux kernel build.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>